added prescriptionDetails model
